Consolidate operand patterns
