+++ src/main/java/edu/kit/compiler/codegen/pattern/Match.java
@@ -5,6 +5,7 @@ import java.util.stream.Stream;
 import firm.nodes.Node;
 
 public interface Match {
+
     public boolean matches();
 
     public Stream<Node> getPredecessors();
