+++ src/main/java/edu/kit/compiler/codegen/pattern/OperandPattern.java
@@ -1,0 +1,75 @@
+package edu.kit.compiler.codegen.pattern;
+
+import java.util.Arrays;
+import java.util.Collections;
+
+import edu.kit.compiler.codegen.NodeRegisters;
+import edu.kit.compiler.codegen.Operand;
+import edu.kit.compiler.codegen.Operand.Immediate;
+import edu.kit.compiler.codegen.Operand.Memory;
+import edu.kit.compiler.codegen.Operand.Register;
+import firm.Mode;
+import firm.bindings.binding_irnode.ir_opcode;
+import firm.nodes.Const;
+import firm.nodes.Node;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+public final class OperandPattern {
+
+    public static Pattern<OperandMatch<Immediate>> immediate() {
+        return new ImmediatePattern();
+    }
+
+    public static Pattern<OperandMatch<Register>> register() {
+        return new RegisterPattern();
+    }
+
+    public static Pattern<OperandMatch<Memory>> memory() {
+        return new MemoryPattern();
+    }
+    
+    @NoArgsConstructor(access = AccessLevel.PRIVATE)
+    public static final class ImmediatePattern implements Pattern<OperandMatch<Immediate>> {
+        @Override
+        public OperandMatch<Immediate> match(Node node, NodeRegisters registers) {
+            if (node.getOpCode() == ir_opcode.iro_Const) {
+                var value = ((Const) node).getTarval();
+                var operand = Operand.immediate(value);
+                return OperandMatch.some(operand, Collections.emptyList());
+            } else {
+                return OperandMatch.none();
+            }
+        }
+    }
+
+    @NoArgsConstructor(access = AccessLevel.PRIVATE)
+    public static final class RegisterPattern implements Pattern<OperandMatch<Register>> {
+        @Override
+        public OperandMatch<Register> match(Node node, NodeRegisters registers) {
+            var register = registers.getRegister(node);
+            if (register >= 0) {
+                var predecessors = Arrays.asList(node);
+                var operand = Operand.register(node.getMode(), register);
+                return OperandMatch.some(operand, predecessors);
+            } else {
+                return OperandMatch.none();
+            }
+        }
+    }
+
+    @NoArgsConstructor(access = AccessLevel.PRIVATE)
+    public static final class MemoryPattern implements Pattern<OperandMatch<Memory>> {
+        @Override
+        public OperandMatch<Memory> match(Node node, NodeRegisters registers) {
+            var register = registers.getRegister(node);
+            if (node.getMode().equals(Mode.getP()) && register >= 0) {
+                var operand = Operand.register(node.getMode(), register);
+                var predecessors = Arrays.asList(node);
+                return OperandMatch.some(Operand.memory(operand), predecessors);
+            } else {
+                return OperandMatch.none();
+            }
+        }
+    }
+}
